Move DependencyMatchingException to the Engine
