+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/geom/geom3d/MeshFeatureTests.java
@@ -43,7 +43,7 @@ import net.imglib2.type.numeric.real.DoubleType;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.scijava.ops.api.features.DependencyMatchingException;
+import org.scijava.ops.engine.DependencyMatchingException;
 import org.scijava.types.Nil;
 
 public class MeshFeatureTests extends AbstractFeatureTest {
